Updated watcher to prevent file locking

+++ watcher.py
@@ -1,33 +1,25 @@
-#!/usr/bin/python
-import time
-import os.path
-
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
 
 from update_file import db, CreateDatabase
 
 
-class MyHandler(FileSystemEventHandler):
+class FoundryHandler(FileSystemEventHandler):
+    def __init__(self, path, file_name):
+        self.file_name = file_name
+        self.observer = Observer()
+        self.observer.schedule(self, path, recursive=False)
+        self.observer.start()
+        self.observer.join()
+
     def on_modified(self, event):
-        # TODO Figure out how to prevent file locking
-        db.create_all()
-        db.drop_all()
-        db.create_all()
-        CreateDatabase().insert_into_database(db)
+        if not event.is_directory and event.src_path.endswith(self.file_name):
+            db.create_all()
+            db.drop_all()
+            db.create_all()
+            CreateDatabase().insert_into_database(db)
 
 
 if __name__ == "__main__":
-    event_handler = MyHandler()
-    observer = Observer()
-    observer.schedule(event_handler,
-                      path="C:/Users/Mikolaj Grobelny/AppData/Local/FoundryVTT/Data/worlds/darklands/data/",
-                      recursive=False)
-    observer.start()
-
-    try:
-        while True:
-            time.sleep(1)
-    except KeyboardInterrupt:
-        observer.stop()
-    observer.join()
+    event_handler = FoundryHandler("C:/Users/Mikolaj Grobelny/AppData/Local/FoundryVTT/Data/worlds/darklands/data/",
+                                   "actors.db")
